refactor: 디렉토리 구조에 맞게 훅 경로 수정 및 새로운 훅 추가
