+++ test/useCalculateOrder.test.tsx
@@ -1,11 +1,11 @@
 import { renderHook } from "@testing-library/react";
 import { describe, it, expect } from "vitest";
-import { useCalculateOrder } from "../src/hooks/useCalculateOrder";
 import { CartItem } from "../src/type/CartItem";
 import {
   FREE_SHIPPING_OVER,
   SHIPPING_FEE,
 } from "../src/constants/priceSetting";
+import { useCalculateOrder } from "../src/hooks/useCalculateOrder";
 
 // 테스트용 목 데이터
 const mockCartItems: CartItem[] = [
